Add more style and some functionality
Added some more styling to the page and achieved the wanted look.
Got the part working where a form pops up when you click the button to
add a new book, but still haven't added the form functionality, only
styled it.

Also made the "read" property of each book change when you click one
of the buttons on the book cards. However this works through code
that uses the id and value of the buttons in order to access the
index of the book, which are assigned through the displayBooks function.
A problem would arise, once removing a book from the library, index of
the books would change but id's and values would not. This would make
pressing a button on one book change read status of another possibly.
Need to correct.